requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
alembic
psycopg2-binary

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# Data Validation & Settings (updated for Python 3.13 compatibility)
pydantic
pydantic-settings
email-validator

# Environment & Configuration
python-dotenv

# File Upload & Storage
boto3
pillow

# Testing
pytest
pytest-asyncio
httpx

# Angel One SmartAPI
smartapi-python
pyotp

logzero
websocket
# Optional Dependencies (uncomment if needed)
# redis==5.0.1
# celery==5.3.6
# stripe==8.6.0
